Rename svg.tsx to rendering.tsx to match the canvas
