+++ src/client/render2d/svg/circle.tsx
@@ -5,7 +5,7 @@ import { Transform } from '../../math/transform'
 import { CircleGeometry } from '../geometry/circle_geometry'
 import { Shape } from '../object/shape'
 
-import { toSvgProps } from './svg'
+import { toSvgProps } from './rendering'
 
 type Props = { model: Shape<CircleGeometry>, world: Transform }
 export const Circle = observer(({ model: { geometry: { x, y, radius }, appearance } }: Props) => (
