+++ src/client/render2d/svg/rendering.tsx
